feature(workspace builder): Add `shell`-configuration to pane layer (#672)
* Added the ability to specify `shell` per pane.

Had to implement the special case for the first pane. In that case the
shell for the window that is created has to be overridden.

This also fixed a bug (in my opinion) that `window_shell` only affected
the first pane, and none of the other panes in a window.

The system unfortunatelly still bailes out, when the specified shell is
not executable (e.g. doesn't even exist). The problem is an unhandeled
exception when the information for the pane can't be extracted (as it
never really started). Not sure how to catch that and provide a sensible
message to the user.

* Added an example for the pane-shell configuration.

* Added the documentation to the example.md file instead.

As the examples.rst files was replaced by the md-files in the main line.