Pass start_directory to new_session when building a session (#829)
This fixes a bug introduced in 2a4714e76b3a85b3391b05413f36623bcb1493f9
(#809) whereby the `start_directory` session config was no longer passed
to the `new_session` method when building a session. This resulted in
the session working directory is not being set correctly, so new
windows/panes created in the session, would not start in the expected
directory.